Create README.md
We at FreeBird Scout are trying to revolutionise the way people plan their trips with a great affordability and comfort. Main aim of our venture is to help users decide where they should visit given the days, budget and type of place. We not only suggest best places around the corner but also the best available stay as well. We are trying to combine travel, stay and dining at one place on an affordable budget while maintaining the quality at its best.